Move message serializatio back to Core
