+++ knightbus-azureservicebus/src/KnightBus.Azure.ServiceBus/ServiceBusConfiguration.cs
@@ -1,5 +1,4 @@
 using KnightBus.Core;
-using KnightBus.Messages;
 
 namespace KnightBus.Azure.ServiceBus
 {
