+++ knightbus-azurestorage/src/KnightBus.Azure.Storage/StorageOptions.cs
@@ -1,5 +1,4 @@
 using KnightBus.Core;
-using KnightBus.Messages;
 
 namespace KnightBus.Azure.Storage
 {
