+++ knightbus-redis/src/KnightBus.Redis/LostMessageBackgroundService.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using KnightBus.Core;
-using KnightBus.Messages;
 using KnightBus.Redis.Messages;
 using StackExchange.Redis;
