+++ knightbus-redis/src/KnightBus.Redis/RedisConfiguration.cs
@@ -1,5 +1,4 @@
 ﻿using KnightBus.Core;
-using KnightBus.Messages;
 
 namespace KnightBus.Redis
 {
