+++ knightbus-redis/src/KnightBus.Redis/RedisManagementClient.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using KnightBus.Core;
-using KnightBus.Messages;
 using KnightBus.Redis.Messages;
 using StackExchange.Redis;
