+++ knightbus-redis/src/KnightBus.Redis/RedisSagaStore.cs
@@ -3,7 +3,6 @@ using System.Threading.Tasks;
 using KnightBus.Core;
 using KnightBus.Core.Sagas;
 using KnightBus.Core.Sagas.Exceptions;
-using KnightBus.Messages;
 using StackExchange.Redis;
 
 namespace KnightBus.Redis
