--- knightbus/src/KnightBus.Messages/IMessageSerialization.cs
@@ -1,10 +1,0 @@
-﻿namespace KnightBus.Messages
-{
-    /// <summary>
-    /// Explicit serialization setting for the <see cref="IMessageMapping"/>
-    /// </summary>
-    public interface IMessageSerialization
-    {
-        IMessageSerializer Serializer { get; }
-    }
-}
