+++ knightbus/tests/KnightBus.Core.Tests.Unit/JsonSerializerTests.cs
@@ -1,4 +1,3 @@
-using System;
 using System.IO;
 using FluentAssertions;
 using NUnit.Framework;
